% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPESDPermissionsListCSV.R
\name{getPESDPermissionsListCSV}
\alias{getPESDPermissionsListCSV}
\title{getPESDPermissionsListCSV}
\usage{
getPESDPermissionsListCSV(name = NULL, section = NULL, unit = NULL)
}
\arguments{
\item{name}{Defaults to \code{NULL}.This is a filter - text entered here will limit the results to only those records containing your
string in the name field (e.g entering "Mik", would return entries with names like "Mike).}

\item{section}{Defaults to \code{NULL}. This is a filter - valid values include \code{"ADMIN"}, \code{"ISAR"}, \code{"SALMON"}, \code{"DADSS"}, \code{"GPSS"},
or \code{"ALL"}.}

\item{unit}{Defaults to \code{NULL}.  This is a filter - text entered here will limit the results to only those records containing your
string in the unit field (e.g. entering "TURT" would return the entries for the turtle unit)}
}
\value{
A data frame
}
\description{
Connects to the PESD Management Team SharePoint site and retrieves an exported csv version of "Staff_Permissions"
list, returning it as a data frame.  The original list lives at
https://086gc.sharepoint.com/sites/msteams_74c888-ManagementTeam/Lists/Staff_Permissions/AllItems.aspx but MS tools
don't allow the extraction of staff names, which exist in the csv. Power Automate is used to extract the list to csv
every day.

This function fetches permission-related details for staff as maintained in the PESD SharePoint Online list. Columns
include employee identifier, organizational unit, mailing lists, employee roles, and permission confirmations. Only
the most relevant metadata columns are returned.
}
\seealso{
\code{\link{getPESDMailingLists}},
  \code{\link{getSharepointSiteMembers}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
