% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePushButtons.R
\name{makePushButtons}
\alias{makePushButtons}
\title{makePushButtons}
\usage{
makePushButtons(arch = NULL)
}
\arguments{
\item{arch}{default is \code{NULL}. The R architecture to use - either 32 or
64 are valid entries. If left as NULL, it will use your default
architecture.}
}
\description{
Generates 6 desktop batch files, one for each grouping category
of \code{autoMakeEmailList()}. Double-clicking a batch file will run the
function with the corresponding \code{by} parameter and copy the result to
the clipboard.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{PESDMgmt}
